// src/core/internal/interfaces/options/ClientOptions.ts
/**
 * @module    core/internal/interfaces/options/ClientOptions
 * @author    Ryuu Mitsuki <https://github.com/mitsuki31>
 * @license   MIT
 * @since     5.0.0
 */

import type { Innertube, Types } from 'youtubei.js';

/**
 * @public
 * @since 5.0.0
 */
export interface ClientOptions {
  /**
   * The maximum number of retries when failed to:
   *
   * - Check the internet connection,
   * - Fetch the video info.
   * - Download the chosen format content.
   *
   * Set this to `0` if you want to disable retries and fail fast.
   *
   * @default 3
   * @see {@link "globals".MAX_RETRIES | MAX_RETRIES}.
   */
  maxRetries?: number;
  /**
   * Whether to enable family safe mode and prevent the session from
   * loading any potentially unsafe content.
   *
   * @default false
   * @see {@link SessionOptions.enable_safety_mode}
  */
  safeMode?: boolean;
  /**
   * An instance of {@link Innertube} represents a YouTube session.
   *
   * @remarks
   * If this option is not set via a configuration file in the CLI application,
   * the global Innertube session will take precedence. If left unspecified,
   * when invoking any core function (such as `getInfo`), it will initiate a new
   * global session with the supplied configuration (`innerTubeConfig`).
   *
   * This will take precedence over the global session if it already exists.
   *
   * @default null
   * @see {@link Innertube.create}
   */
  session?: Innertube | null;
  /**
   * Whether to use the cached video information.
   *
   * If set to `true`, the core function will creates an instance of:
   * - `youtubejs.UniversalCache` : To cache algorithms, session data, and OAuth2 tokens with default persistent cache directory.
   * - `VInfoCache` : To cache the video information separately from Innertube.
   *
   * Default is `true` and cache data generated by **Innertube** will be stored in `"$HOME/.ytmp3-js/.cache/_innertube"`
   * directory (see {@linkcode "utils".INNERTUBE_CACHEDIR | INNERTUBE_CACHEDIR}), unless a
   * `youtubejs.UniversalCache` instance is specified with custom persistent cache directory.
   *
   * For video information cache, the cache will be always stored in `"$HOME/.ytmp3-js/.cache/_vInfoContent"`.
   *
   * @remarks
   * Set this to `false` to disable the all caching behavior including video info cache.
   *
   * @default true
   * @see {@link Types.ICache}
   * @see {@link "cache".VInfoCache}
   */
  useCache?: boolean | Types.ICache;
  /**
   * Whether to skip the internet connection check.
   *
   * The internet check is performed by attempting DNS lookup and will retries
   * based on the `maxRetries` if no internet connection is detected.
   *
   * @default false
   */
  noInternetCheck?: boolean;
}